ItemResponseDatasets.jl

This package provides a quick interface to get different item-response datasets, which could then be fit with IRT models. It could be useful for various types of benchmarking and evaluation.

ItemResponseDatasets.MGKT

ItemResponseDatasets.VocabIQ

ItemResponseDatasets.MGKT.get_marked_df

ItemResponseDatasets.MGKT.get_mgkt

ItemResponseDatasets.VocabIQ.get_marked_df

ItemResponseDatasets.VocabIQ.get_viqt

MGKT

ItemResponseDatasets.MGKT — Module

Taken from the Multifactor General Knowledge Test (MGKT) obtained from OpenPsychometrics.org.

Quoting the README:

"This data was collected through an on-line general knowledge test. Data collection took place 2017-2018. Users were motivated to take the test to obtain personalized results.

This test used a unique question format, where each question was of a type that could have multiple correct answers. In each question, 10 answers were displayed to the respondent and they were told that 5 were correct and to select as many as they knew but not to guess.

The test consisted of 32 questions. They were shown to each subject in a random order."

ItemResponseDatasets.MGKT.get_marked_df — Method

This function gets a marked version of the MGTK.

ItemResponseDatasets.MGKT.get_mgkt — Method

Gets the raw MGKT as a DataFrame.

VocabIQ

ItemResponseDatasets.VocabIQ — Module

Taken from the Vocabulary IQ Test (VocabIQ) obtained from OpenPsychometrics.org.

Quoting the README:

"This data was collected on-line through an interactive test titled "Vocabulary IQ Test" between July 2017 and March 2018.

The main body of the test had 45 vocabulary questions. Each question was a list of five words. Subjects were instructed to select the two on the list that had the same meaning. Subjects were also instructed to not guess and were told there was a -0.35 point penalty for a wrong answer."

ItemResponseDatasets.VocabIQ.get_marked_df — Method

This function gets a marked version of VocabIQ.

ItemResponseDatasets.VocabIQ.get_viqt — Method

Gets the raw VocabIQ as a DataFrame.